1.40
add switch between disk cache & RAM cache, bugfix

+++ sPlayer.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python3
 
-import ctypes, os, platform, sys, time, threading, random, yt_search, simpleaudio
+import ctypes, os, platform, sys, time, threading, random, yt_search, simpleaudio, wave
 from pypresence import Presence
 
 # 사용자 편의를 위해 Youtube Data API v3에 대응하는 API Key를 일부러 깠습니다
@@ -11,6 +11,8 @@ songs = []
 songs_dl = []
 np = []
 qp = False
+hr = False
+hr_enabled = True
 
 RPC = None
 try:
@@ -21,7 +23,7 @@ except:
 
 osis = platform.system()
 
-print("YTSongsBot v1.32")
+print("YTSongsBot v1.40")
 print("type 'h' for help")
 
 def sys_search_video(title):
@@ -29,14 +31,15 @@ def sys_search_video(title):
     print("Search result of keyword " + title)
     print("select a number in search list")
     search_result = yt.search(title, sMax=10, sType=["video"])
-    ni = 0
-    for i in search_result.title:
-        print("#" + str(ni) + " : " + i + " by " + search_result.channelTitle[ni])
-        ni += 1
-    print("x  : cancel current search")
     if qp:
+        print("added song " + search_result.title[0] + " by " + search_result.channelTitle[0] + ", now downloading")
         result = [search_result.title[0], search_result.videoId[0], search_result.channelTitle[0]]
     else:
+        ni = 0
+        for i in search_result.title:
+            print("#" + str(ni) + " : " + i + " by " + search_result.channelTitle[ni])
+            ni += 1
+        print("x  : cancel current search")
         while 1 == 1:
             inpa = input("N>")
             if inpa == "x":
@@ -45,27 +48,27 @@ def sys_search_video(title):
                 try:
                     n = int(inpa)
                     result = [search_result.title[n], search_result.videoId[n], search_result.channelTitle[n]]
+                    print("added song " + search_result.title[n] + " by " + search_result.channelTitle[n] + ", now downloading")
                     break
                 except Exception as e:
                     print(e)
                     continue
-        print("added song " + search_result.title[n] + " by " + search_result.channelTitle[n] + ", now downloading")
         return result
 
 def sys_DownloadSong(vid):
     global osis
     if osis == "Windows":
-        trashbin = ">NUL"
+        os.system('youtube-dl -x --audio-format=wav --output "' + vid + '.%(ext)s" https://www.youtube.com/watch?v=' + vid + ">NUL")
     elif osis == "Linux" or osis == "Darwin":
-        trashbin = ">/dev/null"
+        os.system('youtube-dl -x --audio-format=wav --output "./' + vid + '.%(ext)s" https://www.youtube.com/watch?v=' + vid + ">/dev/null")
     else:
         trashbin = ""
-    os.system('youtube-dl -x --audio-format=wav --output "' + vid + '.%(ext)s" https://www.youtube.com/watch?v=' + vid + trashbin)
 
 def sys_AudioPlayer():
     TitlenoLoop = 0
     global songs_dl
     global np
+    global hr_enabled
     while 1 == 1:
         if not songs_dl:
             if TitlenoLoop >= 1:
@@ -82,19 +85,25 @@ def sys_AudioPlayer():
             TitlenoLoop = 0
             if osis == "Windows":
                 ctypes.windll.kernel32.SetConsoleTitleW("▶ " + np[0])
-            wave_obj = simpleaudio.WaveObject.from_wave_file(np[1] + '.wav')
             if RPC != None:
                 RPC.update(state="노동요 듣는중", details=np[0])
-            os.remove(np[1] + '.wav')
-            play_obj = wave_obj.play()
+            hr_enabled = False
+            if hr:
+                play_obj = simpleaudio.play_buffer(np[3], np[4], np[5], np[6])
+            else:
+                wave_obj = simpleaudio.WaveObject.from_wave_file(np[1] + '.wav')
+                os.remove(np[1] + '.wav')
+                play_obj = wave_obj.play()
             play_obj.wait_done()
             wave_obj = None
             play_obj = None
+            np = None
         time.sleep(.5)
 
 def sys_AutoDownloader():
     global songs
     global songs_dl
+    global hr
     while 1 == 1:
         if not songs:
             pass
@@ -103,7 +112,22 @@ def sys_AutoDownloader():
                 pass
             else:
                 sys_DownloadSong(songs[0][1])
-            songs_dl.append(songs[0])
+            if hr:
+                wave_read = wave.open(songs[0][1] + ".wav", 'rb')
+                audio_data = wave_read.readframes(wave_read.getnframes())
+                num_channels = wave_read.getnchannels()
+                bytes_per_sample = wave_read.getsampwidth()
+                sample_rate = wave_read.getframerate()
+                songs_dl.append([songs[0][0], songs[0][1], songs[0][2], audio_data, num_channels, bytes_per_sample, sample_rate])
+                wave_read.close()
+                wave_read = None
+                audio_data = None
+                num_channels = None
+                bytes_per_sample = None
+                sample_rate = None
+                os.remove(songs[0][1] + '.wav')
+            else:
+                songs_dl.append(songs[0])
             del songs[0]
         time.sleep(.5)
 
@@ -126,7 +150,11 @@ while 1 == 1:
         print("sk           : skip current song")
         print("np           : show what song is playing now")
         print("=====  player options  =====")
-        print("qp           : toggles quick play(always play 1st item in search)")
+        print("qp           : toggles quick play")
+        print("               always play 1st item in search")
+        print("hr           : huge RAM consuming mode")
+        print("               stores downloaded songs to RAM directly")
+        print("               use this command before playing song.")
     elif inp.startswith("r") and inp != "ren":
         if inp == "r":
             print("Usage = r (search keyword)")
@@ -160,6 +188,8 @@ while 1 == 1:
             try:
                 n = int(inp.replace("del ", ""))
                 del songs_dl[n]
+                if not hr:
+                    os.remove(songs_dl[n][1] + '.wav')
             except:
                 print("invalid number")
     elif inp.startswith("pr"):
@@ -171,6 +201,7 @@ while 1 == 1:
                 npr = songs_dl[n]
                 del songs_dl[n]
                 songs_dl.insert(0, npr)
+                npr = None
             except:
                 print("invalid number")
     elif inp == "sk":
@@ -184,5 +215,15 @@ while 1 == 1:
         else:
             qp = True
             print("quick play enabled")
+    elif inp == "hr":
+        if hr_enabled:
+            if hr:
+                hr = False
+                print("huge RAM consuming mode disabled")
+            else:
+                hr = True
+                print("huge RAM consuming mode enabled")
+        else:
+            print("please restart app and use this command before playing song.")
     else:
         continue
